refactor(edge/stacks): separate create by method [EE-4947]
close [EE-4947]

fix(edge/stacks): rename payloads

refactor(edge/stacks): rename methods

refactor(http): move invalid payload error to http

fix(edge/stacks): fix git error
