+++ api/http/errors/invalidpayload.go
@@ -1,5 +1,7 @@
 package errors
 
+import "errors"
+
 type InvalidPayloadError struct {
 	msg string
 }
@@ -10,4 +12,9 @@ func (e *InvalidPayloadError) Error() string {
 
 func NewInvalidPayloadError(msg string) *InvalidPayloadError {
 	return &InvalidPayloadError{msg: msg}
+}
+
+func IsInvalidPayloadError(err error) bool {
+	var payloadError *InvalidPayloadError
+	return errors.As(err, &payloadError)
 }
