--- api/internal/edge/edgestacks/error.go
@@ -1,15 +1,0 @@
-package edgestacks
-
-type InvalidPayloadError struct {
-	msg string
-}
-
-func (e *InvalidPayloadError) Error() string {
-	return e.msg
-}
-
-func NewInvalidPayloadError(errMsg string) *InvalidPayloadError {
-	return &InvalidPayloadError{
-		msg: errMsg,
-	}
-}
